dontstarve/requirements.txt
# 타일 분할 스크립트 의존성
Pillow>=9.0.0
